remove randomize integer dependancy & move Map class in module
